Update and rename (function() {.js to COOKIEAUTO
The provided code is a JavaScript script designed to automate gameplay in a browser-based game called Cookie Clicker. It includes features such as an auto-buyer for purchasing the most expensive available item, an auto-clicker for clicking the cookies, and an auto-golden-cookie-clicker for clicking golden cookies that appear in the game. The script also includes a graphical user interface (GUI) for easy control and configuration of the automation features.

How to Use for Dummies:

Open the Cookie Clicker game in your web browser.
Copy the entire code from the provided script.
Open the developer console of your web browser. You can usually access it by right-clicking on the page, selecting "Inspect" or "Inspect Element," and then navigating to the "Console" tab.
Paste the code into the console and press Enter to run it.
Once the script is running, you will see a GUI displayed on the bottom right corner of the page. The GUI provides the following options:

"Toggle GUI": Click this button to show or hide the GUI panel.
"Start Auto Buyer": Click this button to start the auto-buyer feature, which automatically purchases the most expensive item you can afford.
"Stop Auto Buyer": Click this button to stop the auto-buyer feature.
"Start Auto Clicker": Click this button to start the auto-clicker feature, which automatically clicks the cookies in the game.
"Stop Auto Clicker": Click this button to stop the auto-clicker feature.
"Start Auto Golden Cookie Clicker": Click this button to start the auto-golden-cookie-clicker feature, which automatically clicks on golden cookies when they appear.
"Stop Auto Golden Cookie Clicker": Click this button to stop the auto-golden-cookie-clicker feature.
"Auto Buyer Speed": Use this input field to adjust the speed of the auto-buyer feature (in milliseconds).
"Auto Clicker Speed": Use this input field to adjust the speed of the auto-clicker feature (in milliseconds).
You can customize the auto-buyer and auto-clicker speeds by entering a desired value in the respective input fields. Once you make any changes, the script will automatically use the new speeds.